Remove misleading outdir config
This configuration is not used. Having it here misleads the typescript rookie into thinking that it's required.

The typescript compiler's driver here is the webpack loader. The loader is used to produce intermediate javascript which is then handled by webpack. The output is not directly placed under `outDir` that's specified by the `tsconfig`